+++ src/lib/libbasic/OPEN_D.c
@@ -1,7 +1,7 @@
 
 #include <basic.h>
 
-int OPEN_D(byte f, char* fn, size_t len)
+int OPEN_D(int f, char* fn, size_t len)
 {
     return OPEN('D', f, fn, len);
 }
