+++ src/lib/libbasic/OPEN_I.c
@@ -1,7 +1,7 @@
 
 #include <basic.h>
 
-int OPEN_I(byte f, char* fn)
+int OPEN_I(int f, char* fn)
 {
     return OPEN('I', f, fn, 0);
 }
